+++ src/app/app.routes.ts
@@ -3,23 +3,23 @@ import { Routes } from '@angular/router';
 export const routes: Routes = [
   {
     path: 'home',
-    loadComponent: () => import('./pages/new/home/home.page').then((m) => m.HomePage),
+    loadComponent: () => import('./pages/home/home.page').then((m) => m.HomePage),
   },
   {
     path: 'create',
-    loadComponent: () => import('./pages/new/create/create.page').then((m) => m.CreatePage),
+    loadComponent: () => import('./pages/create/create.page').then((m) => m.CreatePage),
   },
-  {
-    path: 'family',
-    loadComponent: () => import('./pages/tree-families/tree-families.component').then((m) => m.TreeFamiliesComponent),
-  },
-  {
-    path: 'list',
-    loadComponent: () => import('./pages/tree-list/tree-list.component').then((m) => m.TreeListComponent),
-  },
+  // {
+  //   path: 'family',
+  //   loadComponent: () => import('./pages/tree-families/tree-families.component').then((m) => m.TreeFamiliesComponent),
+  // },
+  // {
+  //   path: 'list',
+  //   loadComponent: () => import('./pages/tree-list/tree-list.component').then((m) => m.TreeListComponent),
+  // },
   {
     path: 'view',
-    loadComponent: () => import('./pages/new/view/view.page').then((m) => m.ViewPage),
+    loadComponent: () => import('./pages/view/view.page').then((m) => m.ViewPage),
   },
   // {
   //   path: 'create',
@@ -30,20 +30,20 @@ export const routes: Routes = [
     redirectTo: '/home',
     pathMatch: 'full',
   },
-  {
-    path: 'maps',
-    loadComponent: () => import('./pages/maps/maps.page').then((m) => m.MapsPage),
-  },
+  // {
+  //   path: 'maps',
+  //   loadComponent: () => import('./pages/maps/maps.page').then((m) => m.MapsPage),
+  // },
   {
     path: 'home',
-    loadComponent: () => import('./pages/new/home/home.page').then((m) => m.HomePage),
+    loadComponent: () => import('./pages/home/home.page').then((m) => m.HomePage),
   },
   {
     path: 'create',
-    loadComponent: () => import('./pages/new/create/create.page').then((m) => m.CreatePage),
+    loadComponent: () => import('./pages/create/create.page').then((m) => m.CreatePage),
   },
   {
     path: 'view',
-    loadComponent: () => import('./pages/new/view/view.page').then((m) => m.ViewPage),
+    loadComponent: () => import('./pages/view/view.page').then((m) => m.ViewPage),
   },
 ];
