+++ src/app/pages/create/create.page.scss
@@ -1,31 +1,8 @@
 .form-container {
-  padding: 10px 0 10px 0;
+    padding: 10px 0 10px 0;
 }
-
-.radio-container {
-  overflow: auto;
-  display: flex;
-  height: 50px;
-  align-items: center;
-  gap: 20px;
-}
-
-.image-column {
-  align-items: center;
-}
-
-ion-input,
-ion-select,
-ion-textarea {
-  margin-top: 10px;
-}
-
-.delete-image-button{
-  position: absolute;
-  top: 0;
-  right: 0;
-}
-
-.align-items-center{
-  text-align: center;
+.truncate-text {
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
 }
