--- src/app/pages/maps/maps.page.html
@@ -1,6 +1,0 @@
-<ion-card class="map-card">
-  @if (!isMapLoaded) {
-  <ion-skeleton-text [animated]="true" class="map"></ion-skeleton-text>
-  }
-  <div id="leafletMap" class="map"></div>
-</ion-card>
