--- src/app/pages/maps/maps.page.scss
@@ -1,9 +1,0 @@
-@import "~leaflet/dist/leaflet.css";
-
-.map-card {
-  height: 500px;
-
-  .map {
-    height: 100%;
-  }
-}
