--- src/app/pages/new/create/create.page.html
@@ -1,46 +1,0 @@
-<ion-header [translucent]="true">
-  <ion-toolbar>
-    <ion-buttons slot="start">
-      <ion-back-button></ion-back-button>
-    </ion-buttons>
-    <ion-title> {{ !isEdit ? 'Create' : 'Update' }} </ion-title>
-    <ion-button slot="end" (click)="onSubmit()">
-      <ion-icon name="save" slot="start"></ion-icon>
-      {{ !isEdit ? 'Save' : 'Update' }}
-    </ion-button>
-  </ion-toolbar>
-</ion-header>
-
-<ion-content [fullscreen]="true">
-  <ion-header collapse="condense">
-    <ion-toolbar>
-      <ion-buttons slot="start">
-        <ion-back-button></ion-back-button>
-      </ion-buttons>
-      <ion-title> {{ !isEdit ? 'Create' : 'Update' }} </ion-title>
-      <ion-button slot="end" (click)="onSubmit()">
-        <ion-icon name="save" slot="start"></ion-icon>
-        {{ !isEdit ? 'Save' : 'Update' }}
-      </ion-button>
-    </ion-toolbar>
-  </ion-header>
-
-  <app-select-item
-    [isEdit]="isEdit"
-    [itemFormGroup]="itemFormGroup"
-    [(selectedCategory)]="selectedCategory"
-    [(selectedCategoryItem)]="selectedCategoryItem"
-  ></app-select-item>
-
-  @if(itemFormGroup.valid){
-  <app-item-image
-    [selectedCategoryItem]="selectedCategoryItem()"
-    [(images)]="images"
-  ></app-item-image>
-
-  <!-- Voice Notes -->
-  <app-voice [(audioFiles)]="selectedCategoryItem()!.audioFiles"></app-voice>
-
-  <!-- Map -->
-  <app-item-map [(mapPins)]="mapPins"></app-item-map>
-  }
