--- src/app/pages/new/create/create.page.scss
@@ -1,8 +1,0 @@
-.form-container {
-    padding: 10px 0 10px 0;
-}
-.truncate-text {
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-}
