--- src/app/pages/tree-families/tree-families.component.html
@@ -1,87 +1,0 @@
-<ion-header [translucent]="true">
-  <ion-toolbar>
-    <div class="icon-container">
-      <ion-img src="../../../assets/icon/icon-only.png" class="ion-image"></ion-img>
-      <ion-title> BoomSpoor </ion-title>
-    </div>
-    <ion-buttons slot="end">
-      <ion-button shape="round" [routerLink]="['/create']" color="primary">
-        <ion-icon slot="icon-only" name="add-circle-outline"></ion-icon>
-      </ion-button>
-    </ion-buttons>
-  </ion-toolbar>
-  <ion-toolbar>
-    <ion-searchbar
-      placeholder="Filter"
-      animated="true"
-      [debounce]="250"
-      (ionInput)="filterGroups($event.target.value)"
-    ></ion-searchbar>
-  </ion-toolbar>
-</ion-header>
-
-<ion-content [fullscreen]="true">
-  <ion-header collapse="condense">
-    <ion-toolbar>
-      <ion-title size="large">Tab 1</ion-title>
-    </ion-toolbar>
-  </ion-header>
-
-  <ion-list lines="none" style="height: 100%;">
-    <ion-grid>
-      @if (groups.length > 0) {
-        <ion-row>
-          @for (group of groups; track $index) {
-            <ion-col size="6">
-              <ion-item-sliding style="padding: 5px">
-                <ion-item-options side="end" style="padding: 6px">
-                  <ion-item-option color="primary" (click)="editClicked(group)">
-                    <ion-icon slot="icon-only" name="create" color="dark"></ion-icon>
-                  </ion-item-option>
-                </ion-item-options>
-                <ion-item style="overflow: visible">
-                  <ion-card
-                    class="family-tile"
-                    [attr.id]="group.id['value']"
-                    [routerLink]="['/list', { id: group.id['value'], type: TreeType.Genus }]"
-                    [style.background-color]="pastelColors[$index % pastelColors.length]"
-                  >
-                    <ion-card-header>
-                      <ion-card-title>{{ group.title }}</ion-card-title>
-                    </ion-card-header>
-                  </ion-card>
-                </ion-item>
-              </ion-item-sliding>
-            </ion-col>
-          }
-        </ion-row>
-      } @else {
-        <ion-row>
-          <ion-col size="12">
-            <ion-card class="nursery-tile" style="margin: 0 5px 0 5px" [style.background-color]="pastelColors[0]">
-              <ion-card-header>
-                <ion-card-title>No Families Found</ion-card-title>
-              </ion-card-header>
-            </ion-card>
-          </ion-col>
-        </ion-row>
-      }
-      @if (nursery !== undefined) {
-        <ion-row>
-          <ion-col size="12">
-            <ion-card
-              class="nursery-tile"
-              style="border: 2px solid white; margin: 0 5px 0 5px"
-              [routerLink]="['/list', { id: nursery.id['value'], type: TreeType.Species }]"
-            >
-              <ion-card-header>
-                <ion-card-title>Nursery</ion-card-title>
-                <ion-card-subtitle> All Trees Not Belonging To A Parent </ion-card-subtitle>
-              </ion-card-header>
-            </ion-card>
-          </ion-col>
-        </ion-row>
-      }
-    </ion-grid>
-  </ion-list>
-</ion-content>
