--- src/app/pages/tree-families/tree-families.component.scss
@@ -1,48 +1,0 @@
-.tree-image {
-  display: block;
-  margin-left: auto;
-  margin-right: auto;
-}
-
-.icon-container {
-  display: flex;
-
-  .ion-image {
-    width: 50px;
-    margin-left: 10px;
-    display: block;
-  }
-}
-
-
-.family-tile {
-  height: 110px;
-  width: calc(100% + 16px);
-  border-radius: 3px;
-  margin: 0 -16px 0 0;
-  display: flex;
-  align-items: flex-end;
-  direction: rtl;
-
-  span{
-    height: 78px;
-    width: 100%;
-    margin-bottom: 16px;
-    display: flex;
-    align-items: flex-end;
-    direction: rtl;
-  }
-}
-
-.nursery-tile {
-  height: 100px;
-  margin: 0;
-  display: flex;
-  align-items: flex-end;
-  direction: rtl;
-}
-
-
-ion-item::part(native){
-  padding: 0;
-}
