--- src/app/pages/tree-list/tree-list.component.html
@@ -1,88 +1,0 @@
-<ion-header [translucent]="true">
-  <ion-toolbar>
-    <ion-buttons slot="start">
-      <ion-back-button></ion-back-button>
-    </ion-buttons>
-    <ion-title> {{ title }} </ion-title>
-    <ion-buttons slot="end">
-      <ion-button shape="round" [routerLink]="['/create']" color="primary">
-        <ion-icon slot="icon-only" name="add-circle-outline"></ion-icon>
-      </ion-button>
-    </ion-buttons>
-  </ion-toolbar>
-  <ion-toolbar>
-    <ion-searchbar
-      placeholder="Filter"
-      animated="true"
-      [debounce]="250"
-      (ionInput)="filterTrees($event.target.value)"
-    ></ion-searchbar>
-  </ion-toolbar>
-</ion-header>
-
-<ion-content [fullscreen]="true">
-  <ion-header collapse="condense">
-    <ion-toolbar>
-      <ion-buttons slot="start">
-        <ion-back-button></ion-back-button>
-      </ion-buttons>
-      <ion-title> {{ title }} </ion-title>
-      <ion-buttons slot="end">
-        <ion-button shape="round" [routerLink]="['/create']" color="primary">
-          <ion-icon slot="icon-only" name="add-circle-outline"></ion-icon>
-        </ion-button>
-      </ion-buttons>
-    </ion-toolbar>
-  </ion-header>
-
-  <ion-list lines="none" style="height: 100%;">
-    <ion-grid>
-      <ion-row>
-        @for (tree of treesList; track $index) {
-          <ion-col size="6">
-            <ion-item-sliding style="padding: 5px;">
-              <ion-item-options side="end" style="padding: 6px;">
-                <ion-item-option color="primary" (click)="editClicked(tree)">
-                  <ion-icon slot="icon-only" name="create" color="dark"></ion-icon>
-                </ion-item-option>
-              </ion-item-options>
-
-              <ion-item style="overflow: visible">
-                @if (currentTreeType == TreeType.Species) {
-                  <ion-card
-                    class="tree-image-tile"
-                    [attr.id]="tree.id['value']"
-                    [routerLink]="['/view', { id: tree.id['value'] }]"
-                  >
-                    <img
-                      class="tree-image"
-                      [src]="getImage(tree)?.webviewPath ?? '../../../assets//images/image-not-found.jpg'"
-                    />
-                    <ion-card-header>
-                      <ion-card-title style="font-size: large;">
-                        {{ tree.title }}
-                      </ion-card-title>
-                    </ion-card-header>
-                  </ion-card>
-                } @else {
-                  <ion-card
-                    class="tree-tile"
-                    [style.border]="'3px solid ' + pastelColors[$index]"
-                    [attr.id]="tree.id['value']"
-                    [routerLink]="['/list', { id: tree.id['value'], type: TreeType.Species }]"
-                  >
-                    <ion-card-header>
-                      <ion-card-title style="font-size: large">
-                        {{ tree.title }}
-                      </ion-card-title>
-                    </ion-card-header>
-                  </ion-card>
-                }
-              </ion-item>
-            </ion-item-sliding>
-          </ion-col>
-        }
-      </ion-row>
-    </ion-grid>
-  </ion-list>
-</ion-content>
