--- src/app/pages/tree-list/tree-list.component.scss
@@ -1,29 +1,0 @@
-.tree-image {
-  max-height: 130px;
-  display: block;
-  margin-left: auto;
-  margin-right: auto;
-}
-
-ion-card {
-  min-height: 100%;
-}
-
-.tree-tile {
-  @extend .tree-image-tile;
-  height: 110px !important; 
-  display: flex;
-  align-items: flex-end;
-  direction: rtl;
-}
-
-.tree-image-tile {
-  height: 200px;
-  width: calc(100% + 16px);
-  border-radius: 3px;
-  margin: 0 -16px 0 0;
-}
-
-ion-item::part(native) {
-  padding: 0;
-}
