--- src/app/pages/tree-view/tree-view.component.html
@@ -1,139 +1,0 @@
-<ion-header [translucent]="true">
-  <ion-toolbar>
-    <ion-buttons slot="start">
-      <ion-back-button></ion-back-button>
-    </ion-buttons>
-    <ion-title> {{ tree?.title }} </ion-title>
-    <ion-buttons slot="end">
-      <ion-button fill="clear" (click)="editClicked()">
-        <ion-icon slot="icon-only" name="create"></ion-icon>
-      </ion-button>
-    </ion-buttons>
-  </ion-toolbar>
-</ion-header>
-
-<ion-content [fullscreen]="true">
-  <ion-header collapse="condense">
-    <ion-buttons slot="start">
-      <ion-back-button></ion-back-button>
-    </ion-buttons>
-    <ion-toolbar>
-      <ion-icon name="arrow-back"></ion-icon>
-      <ion-title size="large">{{ tree?.title }}</ion-title>
-    </ion-toolbar>
-    <ion-buttons slot="end">
-      <ion-button fill="clear" (click)="editClicked()">
-        <ion-icon slot="icon-only" name="create"></ion-icon>
-      </ion-button>
-    </ion-buttons>
-  </ion-header>
-
-  <ion-content>
-    <ion-segment [scrollable]="true" value="overview" #infoSegment>
-      <ion-segment-button value="overview">Overview</ion-segment-button>
-      <ion-segment-button value="leaves">Leaves</ion-segment-button>
-      <ion-segment-button value="bark">Bark</ion-segment-button>
-      <ion-segment-button value="fruit">Fruit</ion-segment-button>
-      <ion-segment-button value="flower">Flower</ion-segment-button>
-    </ion-segment>
-
-    <ion-card>
-      @if (infoSegment.value == 'overview') {
-        <ng-container *ngTemplateOutlet="imageSwiper; context: { images: overviewImages }"> </ng-container>
-      } @else if (infoSegment.value === 'leaves') {
-        <ng-container *ngTemplateOutlet="imageSwiper; context: { images: leafImages }"> </ng-container>
-      } @else if (infoSegment.value === 'bark') {
-        <ng-container *ngTemplateOutlet="imageSwiper; context: { images: barkImages }"> </ng-container>
-      } @else if (infoSegment.value === 'fruit') {
-        <ng-container *ngTemplateOutlet="imageSwiper; context: { images: fruitImages }"> </ng-container>
-      } @else if (infoSegment.value === 'flower') {
-        <ng-container *ngTemplateOutlet="imageSwiper; context: { images: flowerImages }"> </ng-container>
-      }
-    </ion-card>
-
-    <ion-card>
-      <ion-card-content>
-        @if (infoSegment.value == 'overview') {
-          <p [innerHTML]="overviewDescription"></p>
-        } @else if (infoSegment.value === 'leaves') {
-          <p [innerHTML]="leafDescription"></p>
-        } @else if (infoSegment.value === 'bark') {
-          <p [innerHTML]="barkDescription"></p>
-        } @else if (infoSegment.value === 'fruit') {
-          <p [innerHTML]="fruitDescription"></p>
-        } @else if (infoSegment.value === 'flower') {
-          <p [innerHTML]="flowerDescription"></p>
-        }
-        <ion-list lines="none">
-          @for (note of tree?.voiceNotes; track $index) {
-            @if (infoSegment.value == 'overview' && note.type === Type.Overview) {
-              <ng-container *ngTemplateOutlet="voiceNote; context: { note: note }"> </ng-container>
-            } @else if (infoSegment.value === 'leaves' && note.type === Type.Leaves) {
-              <ng-container *ngTemplateOutlet="voiceNote; context: { note: note }"> </ng-container>
-            } @else if (infoSegment.value === 'bark' && note.type === Type.Bark) {
-              <ng-container *ngTemplateOutlet="voiceNote; context: { note: note }"> </ng-container>
-            } @else if (infoSegment.value === 'fruit' && note.type === Type.Fruit) {
-              <ng-container *ngTemplateOutlet="voiceNote; context: { note: note }"> </ng-container>
-            } @else if (infoSegment.value === 'flower' && note.type === Type.Flower) {
-              <ng-container *ngTemplateOutlet="voiceNote; context: { note: note }"> </ng-container>
-            }
-          }
-          <ion-item>
-            <ion-toggle [(ngModel)]="showLocation" [enableOnOffLabels]="true">Show Map</ion-toggle>
-          </ion-item>
-        </ion-list>
-      </ion-card-content>
-    </ion-card>
-
-    @if (showLocation) {
-      <app-maps [markerCoordinate]="locations"></app-maps>
-    }
-  </ion-content>
-</ion-content>
-
-<ng-template #voiceNote let-note="note">
-  <ion-item>
-    <ion-button slot="start" fill="clear" (click)="showDeleteVoiceNoteActionSheet(note)">
-      <ion-icon name="trash" slot="icon-only"></ion-icon>
-    </ion-button>
-    <span>{{ note.recordingName }}</span>
-    <ion-button slot="end" fill="clear" (click)="playNote(note)">
-      @if (!note.isPlaying) {
-        <ion-icon name="play" slot="icon-only"></ion-icon>
-      } @else {
-        <ion-icon name="pause" slot="icon-only"></ion-icon>
-      }
-    </ion-button>
-  </ion-item>
-</ng-template>
-
-<ng-template #imageSwiper let-images="images">
-  <swiper-container [pagination]="true" zoom="true" [maxRatio]="8">
-    @for (image of images; track $index) {
-      <swiper-slide>
-        <div class="swiper-zoom-container">
-          @if (image.timestamp !== undefined) {
-            <div class="timestamp">
-              {{ image.timestamp | date: 'yyyy/MM/dd' }}
-            </div>
-          }
-          <img
-            class="tree-image"
-            [src]="image.webviewPath !== '' ? image.webviewPath : '../../../assets/images/image-not-found.jpg'"
-          />
-        </div>
-      </swiper-slide>
-    }
-    @if (images.length === 0) {
-      <swiper-slide>
-        <img class="tree-image" src="../../../assets/images/image-not-found.jpg" />
-      </swiper-slide>
-    }
-  </swiper-container>
-</ng-template>
-
-<ion-popover trigger="click-trigger" triggerAction="click">
-  <ng-template>
-    <app-popover-content></app-popover-content>
-  </ng-template>
-</ion-popover>
