--- src/app/pages/tree-view/tree-view.component.scss
@@ -1,48 +1,0 @@
-// @import 'swiper/css/bundle';
-
-
-.radio-container {
-  overflow: auto;
-  display: flex;
-  height: 50px;
-  align-items: center;
-  gap: 20px;
-  padding: 0 10px;
-}
-
-.tree-image {
-  display: block;
-  margin-left: auto;
-  margin-right: auto;
-  width: 100%;
-  height: auto;
-}
-
-.options-menu {
-  height: fit-content;
-  display: flex;
-  align-items: center;
-  justify-content: left;
-  padding-left: 10px;
-  width: 100%;
-}
-
-.timestamp {
-  position: absolute;
-  top: 5px;
-  left: 5px;
-  width: max-content;
-  height: max-content;
-  background-color: black;
-  color: white;
-  padding: 5px;
-  border-radius: 3px;
-}
-
-.expand-icon {
-  position: absolute;
-  top: 0px;
-  right: 0px;
-  margin: 0;
-  border-radius: 3px;
-}
