+++ src/app/pages/view/view.page.html
@@ -24,7 +24,6 @@
       left: 0;
       width: 100%;
       object-fit: cover;
-      height: calc(100% - 65% + 22px);
     "
     zoom="true"
     navigation="true"
