+++ src/app/pages/view/view.page.scss
@@ -18,7 +18,7 @@ ion-toolbar {
 
 .item-content {
     background-color: transparent;
-    --border-radius: 20px;
+    --border-radius: 15px;
 }
 
 .swiper-zoom-container>img {
