+++ src/app/pages/view/view.page.ts
@@ -1,8 +1,8 @@
 import { CommonModule } from '@angular/common';
-import { Component, CUSTOM_ELEMENTS_SCHEMA, inject, OnInit } from '@angular/core';
+import { Component, CUSTOM_ELEMENTS_SCHEMA, inject, OnInit, ViewChild } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { StatusBar } from '@capacitor/status-bar';
-import { NavController } from '@ionic/angular';
+import { NavController, ViewWillLeave } from '@ionic/angular';
 import { IonButton, IonButtons, IonContent, IonHeader, IonIcon, IonModal } from '@ionic/angular/standalone';
 import { MapComponent } from 'src/app/components/map/map.component';
 import { register } from 'swiper/element/bundle';
@@ -17,10 +17,12 @@ register();
   imports: [IonModal, IonContent, IonHeader, IonIcon, IonButton, IonButtons, MapComponent, CommonModule, FormsModule],
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
 })
-export class ViewPage implements OnInit {
+export class ViewPage implements OnInit, ViewWillLeave {
   private navController = inject(NavController);
 
-  imgHeight: number = 45;
+  @ViewChild(IonModal) modal!: IonModal;
+
+  imgHeight: number = 37;
   notes: string = 'asdasds';
   isModalOpen: boolean = true;
 
@@ -31,11 +33,16 @@ export class ViewPage implements OnInit {
   }
 
   modalChange(event: any): void {
-    this.imgHeight = 105 - +event.detail.breakpoint * 100;
+    this.imgHeight = 101.5 - +event.detail.breakpoint * 100;
   }
 
   async goBack() {
     await StatusBar.setOverlaysWebView({ overlay: false });
     this.navController.navigateBack('home');
+  }
+
+  ionViewWillLeave() {
+    this.modal.dismiss();
+    StatusBar.setOverlaysWebView({ overlay: false });
   }
 }
